+++ src/lib/action/post/delete-post.ts
@@ -6,11 +6,15 @@ import { redirect } from "next/navigation";
 import { prisma } from "@/lib/prisma";
 
 export const deletePost = async (postId: string) => {
-  await prisma.post.delete({
-    where: {
-      id: postId,
-    },
-  });
+  try {
+    await prisma.post.deleteMany({
+      where: {
+        id: postId,
+      },
+    });
+  } catch (error) {
+    throw Error("投稿の削除に失敗しました");
+  }
 
   revalidatePath("/");
   redirect("/");
